Fix ESTIMATE_COST JavaScript UDF parameter types
Error: Language JAVASCRIPT does not support type 'NUMBER(38,0)' for argument type

Per Snowflake documentation, JavaScript UDFs do not support INTEGER type.
JavaScript has only 'number' type (double/float). Changed INTEGER parameters
to FLOAT for JavaScript UDF compatibility.

Fixed: Changed input_tokens and output_tokens from INTEGER to FLOAT

+++ setup/07_rcm_native_agent_production.sql
@@ -152,8 +152,8 @@ $$;
 
 -- Calculate cost estimate
 CREATE OR REPLACE FUNCTION ESTIMATE_COST(
-    input_tokens INTEGER,
-    output_tokens INTEGER,
+    input_tokens FLOAT,
+    output_tokens FLOAT,
     model STRING
 )
 RETURNS FLOAT
